move type_domain to separate files
