+++ src/crab/dsl_syntax.hpp
@@ -31,6 +31,12 @@ inline linear_expression_t operator-(const number_t& n, const linear_expression_
 
 inline linear_expression_t operator-(const linear_expression_t& e, const number_t& n) { return e.subtract(n); }
 
+inline linear_constraint_t eq(const variable_t a, const variable_t b) {
+    return {a - b, constraint_kind_t::EQUALS_ZERO};
+}
+
+inline linear_constraint_t neq(const variable_t a, const variable_t b) { return {a - b, constraint_kind_t::NOT_ZERO}; }
+
 inline linear_constraint_t operator<=(const linear_expression_t& e1, const linear_expression_t& e2) {
     return linear_constraint_t(e1 - e2, constraint_kind_t::LESS_THAN_OR_EQUALS_ZERO);
 }
